streamDM-103: updates to rf and ht to address comments in PR
